Moved the models to a package called entity in the local package
